镭射点单独管理，脱离枪械矩阵

+++ src/main/java/com/modularwarfare/client/fpp/basic/configs/AttachmentRenderConfig.java
@@ -84,7 +84,7 @@ public class AttachmentRenderConfig {
     public static class Laser {
         public float[] laserColor = {1.0f, 0.0f, 0.0f};
         public float laserAlpha = 1.0f;
-        public float maxDistance = 100f;
+        public double maxDistance = 100f;
         public float dotSize = 2.0f;
     }
